сделал модуль 8 часть 1
