+++ js/main.js
@@ -1,13 +1,19 @@
 import {makeOffers} from './data.js';
+import './render-form.js';
+import {getCardElement} from './render-form.js';
+import {hideForm} from './action-form.js';
+import {showForm} from './action-form.js';
+import {hideFilter} from './action-filter.js';
+import {showFilter} from './action-filter.js';
 import './form.js';
-import {getCardElement} from './form.js';
-import {getHideForm} from './action.js';
-import {getShowForm} from './action.js';
 
 const offers = makeOffers();
 
 const map = document.querySelector('.map__canvas');
 map.append(getCardElement(offers[0]));
 
-getHideForm();
-getShowForm();
+hideForm();
+showForm();
+hideFilter();
+showFilter();
+
